TogglePOVKK - Proper init and refactor
